requirements.txt
fastapi
nltk
PyPDF2
Spire.Doc
python-docx
psycopg2-binary
uvicorn
python-multipart
sqlalchemy
asyncpg
databases
